python/api-python/requirements.txt
fastapi
uvicorn[standard]
pydantic
render-sdk>=0.5.0
python-dotenv
requests
beautifulsoup4
httpx
slowapi
boto3>=1.34.0
ruff>=0.8.0